% Auto-generated: do not edit by hand
\name{''DragResize}

\alias{''DragResize}

\title{DragResize component}

\description{

}

\usage{
''DragResize(id=NULL, dragGrid=NULL, onDragStop=NULL, onResizeStop=NULL,
position=NULL, resizeGrid=NULL, setState=NULL, size=NULL)
}

\arguments{
\item{id}{Character. }

\item{dragGrid}{Named list. }

\item{onDragStop}{}

\item{onResizeStop}{}

\item{position}{Named list. }

\item{resizeGrid}{Named list. }

\item{setState}{}

\item{size}{Named list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

