% Auto-generated: do not edit by hand
\name{''DragResize}

\alias{''DragResize}

\title{DragResize component}

\description{

}

\usage{
''DragResize(children=NULL, id=NULL, h=NULL, w=NULL, x=NULL, y=NULL)
}

\arguments{
\item{children}{Named list. }

\item{id}{Character. }

\item{h}{Character. }

\item{w}{Character. }

\item{x}{Numeric. }

\item{y}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

