% Auto-generated: do not edit by hand
\name{''DragResize}

\alias{''DragResize}

\title{DragResize component}

\description{

}

\usage{
''DragResize(children=NULL, id=NULL, bounds=NULL, className=NULL,
disableDragging=NULL, dragGrid=NULL, enableResizing=NULL,
h=NULL, resizeGrid=NULL, style=NULL, w=NULL, x=NULL, y=NULL)
}

\arguments{
\item{children}{Named list. }

\item{id}{Character. }

\item{bounds}{Character. }

\item{className}{Character. }

\item{disableDragging}{Logical. }

\item{dragGrid}{Unnamed list. }

\item{enableResizing}{Logical. }

\item{h}{Character. }

\item{resizeGrid}{Unnamed list. }

\item{style}{Named list. }

\item{w}{Character. }

\item{x}{Numeric. }

\item{y}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

