% Auto-generated: do not edit by hand
\name{''DragResize}

\alias{''DragResize}

\title{DragResize component}

\description{

}

\usage{
''DragResize(id=NULL, dragGrid=NULL, height=NULL, onDragStop=NULL,
onResizeStop=NULL, position=NULL, resizeGrid=NULL,
size=NULL, width=NULL, x=NULL, y=NULL)
}

\arguments{
\item{id}{Character. }

\item{dragGrid}{Named list. }

\item{height}{Numeric. }

\item{onDragStop}{}

\item{onResizeStop}{}

\item{position}{Named list. }

\item{resizeGrid}{Named list. }

\item{size}{Named list. }

\item{width}{Numeric. }

\item{x}{Numeric. }

\item{y}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

